refactor(types): consolidate routing declarations
