+++ types/server-options.d.ts
@@ -12,8 +12,7 @@ import { FastifyInstance } from './instance'
 import { FastifyBaseLogger, FastifyChildLoggerFactory, FastifyLoggerOptions, LogController, PinoLoggerOptions } from './logger'
 import { FastifyReplyForRoute } from './reply'
 import { FastifyRequestForRoute } from './request'
-import { FastifyRouterOptions, FindMyWayVersion } from './router-options'
-import { RouteGenericInterface } from './route'
+import { FastifyRouterOptions, FindMyWayVersion, RouteGenericInterface } from './route'
 import { FastifySchema, SchemaErrorFormatter } from './schema'
 import { FastifyServerFactory } from './server-factory'
 import { FastifyTypeProvider, FastifyTypeProviderDefault } from './type-provider'
